Basic image reading beta

+++ include/console_bmp/bit_view.hpp
@@ -7,7 +7,7 @@ namespace console_bmp {
 
 struct BitView {
     uint8_t* data_ptr;
-    uint8_t bits_offset;
+    size_t bits_offset;
 
     void read_into(size_t size_bits, size_t offset_bits, void* out_data) const;
     void write_from(size_t size_bits, size_t offset_bits, const void* in_data) const;
